データ保存まで

+++ database/migrations/2023_07_26_081245_create_products_table.php
@@ -14,12 +14,12 @@ class CreateProductsTable extends Migration
     public function up()
     {
         Schema::create('products', function (Blueprint $table) {
-            $table->unsignedBigInteger("id");
-            $table->integer("user_id");
+            $table->bigIncrements("id");
+            $table->unsignedBigInteger("user_id");
             $table->string("product_name");
             $table->integer("stock");
             $table->integer("price");
-            $table->integer("company_id");
+            $table->unsignedBigInteger("company_id");
 
             $table->timestamps();
         });
